npm: inline native staging helper
